refact: save emotes in global context
